add app bar and refresh controll

+++ lib/features/weather/domain/entities/weather_forecast_entity/weather_forecast_minutly_entity.dart
@@ -6,7 +6,7 @@ class WeatherForecastMinutlyEntity extends Equatable {
   final int dt;
 
   /// Precipitation, mm/h. Please note that only mm/h as units of measurement are available for this parameter
-  final int precipitation;
+  final double precipitation;
 
   const WeatherForecastMinutlyEntity({
     required this.dt,
